server/requirements.txt
langchain
chromadb
huggingface-hub
sentence-transformers
pymupdf
langchain_ollama
numpy
fastapi[standard]
sqlmodel
python-dotenv
psycopg2-binary>=2.9